% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Server logic for the HangmanneR app}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, output, session}{Objects supplied by Shiny.}
}
\value{
Called for its side effects on \code{output}.
}
\description{
Server logic for the HangmanneR app
}
