% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{User interface for the HangmanneR app}
\usage{
app_ui()
}
\value{
A \code{\link[shiny:fluidPage]{shiny::fluidPage()}} with the game, settings and about tabs.
}
\description{
User interface for the HangmanneR app
}
