% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HangmanneR-package.R
\docType{package}
\name{HangmanneR-package}
\alias{HangmanneR}
\alias{HangmanneR-package}
\title{HangmanneR: Play Hangman with CRAN Package Names}
\description{
A Shiny game of hangman played with CRAN package names. Letters are guessed one at a time to uncover a package sampled from a snapshot of CRAN metadata that ships with the package. Difficulty is set by the length of the package name, play can be restricted to packages authored by Posit Software, PBC, and the package description is available as a hint with the name redacted.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matiasandina/HangmanneR}
  \item Report bugs at \url{https://github.com/matiasandina/HangmanneR/issues}
}

}
\author{
\strong{Maintainer}: Matias Andina \email{matiasandina@gmail.com} [copyright holder]

}
\keyword{internal}
